+++ src/main/java/likelion/festival/converter/NotificationStringToRequestDtoConverter.java
@@ -3,7 +3,7 @@ package likelion.festival.converter;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import likelion.festival.dto.NotificationDto;
+import likelion.festival.notification.application.dto.NotificationDto;
 import org.springframework.core.convert.converter.Converter;
 import org.springframework.stereotype.Component;
