+++ src/main/java/likelion/festival/entity/Image.java
@@ -2,6 +2,7 @@ package likelion.festival.entity;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import likelion.festival.booth.domain.Booth;
+import likelion.festival.notification.domain.Notification;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
