+++ src/main/java/likelion/festival/notification/domain/repository/NotificationRepository.java
@@ -1,11 +1,10 @@
-package likelion.festival.repository;
+package likelion.festival.notification.domain.repository;
 
-import likelion.festival.entity.Notification;
-import likelion.festival.entity.NotificationType;
+import likelion.festival.notification.domain.Notification;
+import likelion.festival.notification.domain.NotificationType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.ArrayList;
-import java.util.List;
 
 public interface NotificationRepository extends JpaRepository<Notification,Long> {
