+++ src/main/java/likelion/festival/service/ImageService.java
@@ -2,7 +2,7 @@ package likelion.festival.service;
 
 import likelion.festival.dto.ImageDto;
 import likelion.festival.booth.domain.Booth;
-import likelion.festival.entity.Notification;
+import likelion.festival.notification.domain.Notification;
 import likelion.festival.repository.ImageRepository;
 import likelion.festival.util.MD5Generator;
 import lombok.RequiredArgsConstructor;
